Added Projection capabilities to the runtime
Projection is the problem of answering a query after hypothetically
executing a sequence of updates.  An admin might pose the following
problem: add this logical port, then this logical router, connect
them this way, and then tell me what all the errors are.

This change adds a Projection call to the runtime.  The sequence
of updates can be action invocations or rule/atom insert/deletes.

The current implementation returns the set of all query instances
true in the resulting state.  Relatively simple to write a wrapper
that computes the set of *changes* in the query resulting from
the update sequence as well.

Issue: #
Change-Id: I69985df9dad6ae920be129191dd3bd70fa4f93dc