Adding Neutron action descriptions
Adding realistic (though obviously not perfect)
action descriptions of the core Neutron API.

Networks, subnets, ports are done.  Outstanding issues
are listed at the top of examples/neutron.action.  Some of these
need to be addressed, but the goal of this was to understand
the extent to which Neutron/Nova/etc. actions could be expressed
and reasoned about.  Plan is to revisit the list of issues
after a similar investigation into Nova.

- Added rudimentary tests for networks, subnets, ports.

- Also fixed a bug for theory inclusion in top_down_evaluation.

- Made minor improvements to the code (e.g. there is now a Theory obj
that all the theories inherit from).

- Increased syntax to include several atoms in the head of each rule.
Necessary to describe update sequence for testing Neutron API.

All tests pass

Issue: #
Change-Id: I63b188acbe51b07e0ceffc9c747b5b62c46c92bf