Added registry for Datalog builtins
Previously, all tables in Datalog had to be defined by the user.

This change begins adding builtins to Datalog: tables that are computed via
Python functions, e.g. arithmetic.  This change provides
a registry for builtins and seeds that registry.  A future
change will connect that registry to the Datalog runtime.

Change-Id: I04cb940cb1ff6773e42c1b382a1675ef3602518b