Robustness improvements to NeutronDriver
Previously we were generating each row of a Neutron
network table by iterating over the keys of a dict.
That iteration is not guaranteed to occur in the same order
every time, causing fields in the network table to
appear in different orders.

This change makes the field in the network table
a fixed integer for each of the dictionary keys
returned by the Neutron client.

Change-Id: I873b1652b2da12a6fbac9fea9c295dfa40b4a8bb