Added API data models
The API system passes HTTP(S) requests down to data models, which
in turn are responsible for producing (Python) answers to the API
calls.

This change adds API models for policy, datasource, table, row, rule.
Not every model is as flushed out as we would like, but there is
enough functionality to do end-to-end testing.

Change-Id: Ia3c66262d8ad94f419d3b2174d20c92a45648215