Fix relative import path in source tree
This patch fixes the relative import path to use congress
as the base to be consistant with other openstack projects. In addition,
this commit removes the the run_congress_server script and replaces that
with bin/congress-server and the run_test script.

Unit tests can now be successfully be run by issuing:
make
tox -epy27 (once the tox.ini enables tests again).

Closes-bug: 1338840

Change-Id: I371192e9573a7f1ca4ea8e1d6a707e265e8f8dc9