Refactor congress eventlet server code
This patch removes the current wsgi-server in place of the one that keystone
uses which is basically identical to the one that was already in congress.
The only reason I switched it out for the keystone one is just so copying
the keystone integration out of keystone to congress would be easier.
In addition, this patch adds mutiworker support so can use multiple processes
to run our api server (though currently we just default that to 1).
This patch also includes test cases for the eventlet server.

Change-Id: I25ab083d359a4bd85c93356e5261d31061a7e9aa
Closes-bug: 1339880