Avoid redundant datasource tuple translation
Every time Neutron is queried for its
current information, the Neutron driver
converts that information to tuples, even
if the data is exactly the same as it was
previously.

This change only runs the conversion to
tuples when the data is different than
it was previously.  This has the benefit
that bug #1344475 has less impact
in the common case where data has not changed.

Related-bug: 1344462
Change-Id: I53132258804eeff8020f07d46dc2515aa23ee368