Fixed bug with auto-generated tables
Sometimes the policy runtime creates new tables from existing table
names.  When it does so for a tablename that references an external
datasource, e.g. neutron:ports, it leaves the datasource prefix
in tact.  This causes the DSE incarnation of the policy engine
to try to create a data source with a non-existent data source
name.

This fix checks if the table-name is auto-generated before
trying to spin up a datasource service.

Change-Id: Ic3995c888c74640f9b02a2acb986144b9cc536db