Add devstack integration files
This patch introduces integration files that we need in order to
use devstack to configure congress automatically for us. This is
an important step as we need this automation to integrate CI testing.

This patch also changes the .gitignore file as by default the path
in here is match globally in the project thus lib in .gitignore was
preventing me from adding the contrib/devstack/lib/congress file.

closes-bug: 1343545
Change-Id: I03613270cd1c8b5910499e74398a39a5167bae69