Made policy_path optional
Previously, when launching Congress we needed to give an absolute
pathname to the chunk of policies we wanted initially loaded.
One of those policies includes configuration data about the datasource
drivers, without which Congress does not run.

This change ensures that if a policy_path is not provided,
the default is congress/etc/snapshot.

Change-Id: Ie8614fc984f14b1f8a1a3b8280581bf4d54a48f7