Simplify prepare-devstack script
Previously one was required to clone the congress source tree then
run prepare-devstack.sh and then stack.sh. After stack.sh completed one
might be left with two congress directories depending if they put it
in $DIST (/opt/stack) which could require them creating this directory and
chowning it to themselves. In addition, it would require you to install
realpath as a dependency. Instead of doing this we've updated this script to
just download the required files without haveing to clone the initial congress
repo.

Change-Id: I221921c656a0425c4611878a165d65305d0c60d2
closes-bug: 1356022