Added per-datasource configuration
Previously, all datasources had the same username/password/tenant
configuration information.  Additionally, the name to module mapping
was stored in a different place (but each data source could have
a different module).

This change puts all of the configuration information for datasources
into a single file, and enables the configuration information to
be different for each datasource.

Closes-bug: #1356615
Change-Id: I0fc4274c452ff91663eded7828dc8d8679935b4e