Addendum to docs
Previously, the docs did not show how to configure data-sources,
and they relegated the API to the design doc.

This change adds the missing information to the documentation.
(Always having the docs reflect the *current* API offerings is
useful, and frees up the design doc to reflect current and
future offerings.)

Change-Id: Id84bfb0be865b9d5d5ed60cda75bfc1603fdcf0f