Fixed Neutron translation of external_gateway_info
Previously, the translation of the router's external_gateway_info
basically either returned True or False, indicating whether or
not the external_gateway_info was None or not.

This change properly translates external_gateway_info
into either None or a hash table.  It also adds tests
for router and security_group translations.

Change-Id: Ie482d57d6e644bffc9cf0f57a578f04a6e5fabbe