Default setting for verbose/debug should be True
There was a patch that was merged that changed these values to False.
With these values being false it makes it hard for one to track down why
things are failing without enabling it. One can change this setting
to False in their own setup or set: ENABLE_DEBUG_LOG_LEVEL=False in
ones localrc file to turn of debugging.

Change-Id: I816fa767948fd3f7c5f6939625f844fb3a4440a1
Closes-bug: 1364130