Fix missing imports causing gate failures
We have tests that import fakes from the python-novaclient. For example,
in congresss/datasources/tests/unit/fakes.py. We should not be doing this
as the novaclients tests are not intended to be a stable interface thus
because additional imports were added to the novaclient file here this
broke us since we don't require these imports. This patch is a quick
hack to get the unit tests working though we need to fix this otherwise
our tests will mostly break again in the future when the novaclient changes.

Change-Id: Iafceea05a3989012b460e48ebfa2021e8d4e74c5
Closes-bug: 1364136