Fixed bug in materializedViewTheory
One of the recently reactivated tests was broken,
probably because of some code cleanup done when
the tests were deactivated.  The bug arises with
how the materializedViewTheory handles the insertion
or deletion of rules when data is already present.
The bug arose because we were trying to compute
the new data that should be added/deleted when
inserting/deleting a rule at the time that
rule was enqueued to the processing queue.

This change fixes that bug by moving
the computation of new data from the enqueueing
step to the processing step.  It also reactivates
the test that was broken.

Change-Id: I690d48c4e1f99a214d7000741e7476ada90d9e3e