Remove unneed step to set fixed-ip and neutron cleanup
This patch removes an uneeded step to set a fixed ip on a port. This is
unneeded because neutron automatically allocates you a fixed-ip when
the port is created. In addition, this hides the provider information which is
only shown if one configures neutron to use vxlan/vlan for the transport type.

Change-Id: Id08d456cac0dcbf9b8517779ebaccfd5d9025bef