Fixed bug in config file poll_time parameter
Previously, the poll_time of each datasource could not be set
in the config file because the config read the value in as a string,
and the datasource driver didn't convert it to an int.   Thus setting
the poll_time in the config file caused an error in the driver.

This change casts the poll_time to an int on initialization.

Change-Id: I0fe1db536f7e91948a59c6f80a87baa896fa7fcd