Don't use keystone client before initializing it
Eventlet was calling poll() before the KeystoneDriver finished, so the the
keystone client object was not initialized yet.  This caused an AttributeError
in KeystoneDriver.  This change separates the KeystoneDriver constructor from
initialization of the keystone client object.  This separation ensures that
the first time eventlet calls update_from_datasource(), the KeystoneDriver is
fully initialized.

Change-Id: I55f76d5dc115920ea04d0a954919d9ef329fb886