Fix bug in value_to_congress
The previous version checks if a value is equal to either True or False, if
so, it converts the value to a string.  The problem is that 0 and 1 are equal
to False and True respectively in python which would cause integer values 0
and 1 to be converted to '0' and '1'.

This change uses isinstance(obj, bool) and does so before isinstance(obj, int)
because a bool value is also an instance of int.  Also, add a test to verify
this bug.

Change-Id: Ifbf17378c1ca660c47421b11a30204f9be1af056