Added delta version of simulate
Previously, the Runtime.simulate functionality would only return
the actual query contents that you specified.  But sometimes
the user wants to only know the difference in that query
caused by the updates that are being simulated.

This change adds a parameter to Runtime.simulate that causes
it to return the difference in the query instead of the full
contents of the query.

Change-Id: I265ea408495479803c4f9cb6be5e6002c25b8776