Refactor module_schemas in policy engine Runtime
Previously, the schema for each module was stored
separately from theories.  A single theory could
define tables from any module.

This change is one of several that ensures
a theory and module are interchangeable terms.
All the tables defined for module M are defined
within theory M.   As such, the term 'module' has
been replaced with 'theory', and schemas are now
a feature of a theory.

Implements-blueprint: multiple-policies
Change-Id: Ieee9dd9b90c4726bfea65c02a8dd0f9db59ba7e9