Increase timeout on retry decorator
Some unit tests are still failing due to not waiting long enough. This
is caused because we are retrying every 10ms. This patch sets the retry
interval to 100ms second instead to avoid this failure. In the future we'll
refactor these tests to not test things in an undetermistic way.

Change-Id: I95e67ebdb59724d6f7c4bfe6a2970879ee38e6fb
Closes-bug: 1392561