Move inline funtions out of convert_obj
This patch pulls the inline methods out of convert_obj to be private methods
in the DataSourceDriver in an attempt to improved readability. Inline functions
are generally dangerous to use if they are long, in addition one needs to
gernally take special care in variable scoping from the outter function which
can be error prone. This similar clean up should also be done in other places
but this patch just handles convert_obj for now.

Change-Id: I707cd5005400e4bb76e9adf57096db0da6f504e4