Refactor get_credentials out of DataSourceDriver class
This patch factors out the method get_credentials() out of the
DataSourceDriver class. In addition, removes the logic that allows
the tests to pass in a client() to a datasource to avoid mocking out the
client import. The motiviation for me to remove this is we can use mock
directly to do this without having to have additional code logic in a
datasource driver to test it.

Partially implements blueprint refactor-datasource-driver-framework

Change-Id: I7b07ff1fc34eff9aa28d310c8482edfdf9fb4e92