Fix bug with column-references
There were a number of places where we were calling compile.parse
instead of runtime.Runtime.parse.  The latter is aware of schemas
but the former is not.  This leads to bugs when column-references
are used in rules because column-references are resolved at parse
time.

This change is a search-and-replace over all the files in the repo
to use runtime.Runtime.parse instead of compile.parse.  All occurrences
except in test-code were replaced.  Occurrences in test-code that
instantiated runtime.Runtime were replaced.  Occurrences in test-code
that did not instantiate runtime.Runtime were not replaced since
they did not utilize column references, except in those cases where
they passed in the requisite schemas into compile.parse.

Change-Id: I6566ba06d7b5371b04220a8af747c7065feefc55