Fix for simulation bug
Previously, simulation did not properly insert data into
policies other than the one given for the query.  This led
to incorrect results.  For example nova:p+(1) did not get inserted
into the nova policy, hence a query relying on nova:p(1) could
fail.

This change identifies the policy to which a data element is targeted
and inserts/deletes it from the appropriate policy.

Change-Id: I43a2deb346d0b2a09e59398607cd5d545b711ecc
Closes-bug: 1400921