Fix a transient test failure in tempest tests
The previous version of these tests fetch data directly from the data service
and from congress and compares the two responses.  However, the test is not
synchronized with whatever inserts data into the data service, so when the
test starts, the data service might not have all the data.  While the test is
running, the data service can get more data and give the data to congress.
This causes the test to think congress has data that the data service does
not, leading to a KeyError.

This changes makes the test fetch data from the data service and congress each
time the test tries to compare the two.  This ensures that eventually, the two
datasets converge.  It also avoids the KeyError when the congress data has
something the data service does not.

Change-Id: Ie5e6c82ae2777921d1c6367e5297351455f69683