Speed up dependency graph computation
Previously, every insert or delete into any policy resulted
in re-computation of that policy's dependency graph and
a check that no recursion occurred either within a policy
or across policies.  This dependency graph computation was
expensive because (i) instead of updating the graph based
on the policy statements inserted/deleted, the graph
was recomputed from scratch and (ii) each policy had
its own dependency graph, and to check for recursion
across policies we computed the union of those individual
graphs.  In addition, we were copying the list of current
rules before doing any of that computation.

This change eliminates all 3 sources of inefficiency.
There is now a single, global dependency graph; that graph
is updated incrementally on each insert/delete; there is
no copying of rules.  In addition, we ensure that the
dependency graph computation only happens for non-atoms
(since inserting atoms is a common case and atoms cannot
contribute to recursion).

The implementation adds a new graph data structure (BagGraph) that
makes the incremental computation of dependencies easy
and fast.  This change adds direct tests for the original Graph
data structure as well as for the new BagGraph.

Change-Id: I5f72b3c6193f5d35ac293f353b6d8dcf13dd639d