Introduce test_datasource_driver_config for all openstack drivers
This patch pulls out the testcase TestDataSourceDriverConfig from the neutron
driver to it's own testcase that run against all of the openstack based
drivers (expect swift as it s does not to use the openstack_args_function).
This change buys us additional coverage on all drivers and is a step
forward in removing the neutron datasource_driver.

Closes-bug: 1410938

Change-Id: If32a30320b5ef67895371ab48c61d61bb586e72a