Fix multipolicy simulation bug
Previously, there was a bug in the simulation functionality when
using action descriptions.  The simulation functionality simply failed
to compute the proper updates given an action instance, thereby the results
of simulation were wrong.

This bug fix does a number of things
1) Only allows policies to appear in the tablenames in the head of a rule
within an ActionTheory if that tablename is an update (i.e. has a + or -
at the end of the tablename).

2) Changes the policy evaluation algorithms to treat an update literal
as a local table even if the literal references a separate policy.

3) Changes the implementation of simulation (the consequences
routine in particular) so that it would generate the appropriate
queries for computing proper updates caused by an action instance.

4) Changes the routine for get_arity to take a theory as an argument
and enable a local_only option so that we compute the arity from
the proper policy in the case of updates.

Change-Id: I27564891e0ea4cc463127f1811edd275c27d568f
Closes-bug: 1413793