Move self.rules implementations to TopDownTheory subclasses
Previously, TopDownTheory assumed that it's subclasses would define a field
called self.rules.  This change eliminates that assumption by moving any code
that uses self.rules to the subclass NonrecursiveRuleTheory.

This new organization makes it cleaner for a subclass of TopDownTheory to
implement self.rules in a different way because now TopDownTheory will not be
reaching into self.rules at all.  Instead, methods like head_index(), head(),
and body() are all implemented in the subclass.

Change-Id: Id0c2ff09a11ca37dfafda6efda951d503d2b05b9