Cache schema generation
Previously each time one called get_schema() on a datasource_driver we would
generate it by walking through all of the translators. This patch changes the
logic so that we store the schema when we register a datasource so we do not
have to generate it each time. Note: the schema never changes durring run time
so we're able to cache it like this.

Change-Id: I4deead9fa698e681c0f0a8f07a4ef7d95726bae7
Closes-bug: #1415208