Make datasources configurable via the API
This patch allows datasources to be configured via the API. It exposes a new
api call /v1/drivers/<driver> which returns the configuration
needed by a driver in order to configure it. Then, a user will make a post to
/v1/data-sources with the config to create the datasource.

This patch  makes a few structural changes to congress the most noted that it
makes d6Cage a singleton which is needed to be able to dynamically add new
datasources to the message bus.

Also, lots of the API code was changed in order to start moving toward an API
that supports multi tenancy which should now be easy to add.

I apologize that this patch set is so big though the scope of changes needed
to be down to support a configurable API were large and through out many
places in the code. Several cleanup patches will be push on top of this which
should be easier to review.

Implements blueprint: api-configurable-datasources

Change-Id: If82b505e1835254216e19eb1c55a035d2c0b8a3e