Add DSE/Runtime/Datasource performance tests
These tests check how long it takes for a datasource driver to translate
Python data into tables, send the tables across the DSE to the policy
engine, and have the policy engine load the tables.  There is a separate
test that circumvents the datasource driver translation process and
just checks the time for sending tables across the DSE and initializing
them in the policy engine.

Change-Id: I9715eb5d80a1ec48722f3f36c7ececf9714047cd