Fix index lookup in runtime joins
The previous code has a bug: when the runtime joins a key against a table
using an index, if there are no matching rows for that key, the runtime ends
up iterating over all rows in the table.

This change fixes the bug.  If there is no matchh for the key, then the join
can immediately return the empty set.  There is no need to iterate through all
the rows in the table.

Also add a test for this bug.

Change-Id: Ic6e1899ec393c21aba745a24be9d476ded21a8b1