Fix failing test case TestDsePerformance
This testcase TestDsePerformance did not set the db url so the tests would
sometimes fail (not sure why they didn't always fail). Anyways, this patch
fixes this issue by setting the config option in the base class that
deals with tests that use the db. Moving this here ensures no one needs to
call this method in testcases which touch the db which is better anyways.

Closes-bug: 1418188

Change-Id: Ibc6121f6671af62498d2baa23c14e60aef9b356d