API: Improve status api code and add tests
The status API was setup as a CollectionHandler though the API request
was structured to be handled by an ElementHandler this get_items() was
being called instead of get_item(). In addition, nothing was setup to
ever call get_item. This patch corrects the structure and configures it
to be an ElementHandler.

In additon, this patch changes the status api so that responses are now
returned in this format: {'last_updated': 'now', 'last_error': 'foo'}
previously: {'results': [{'last_updated': 'now'}, {'last_error': 'foo'}]}

Lastly, this patch adds code coverage for this api.

Change-Id: Icb245cf9f21d8cef3cc45a47df98038c760c9770