Add analysis routine for subpolicy construction
This change adds find_subpolicy to the Runtime engine, which analyzes
a given policy to carve out the subpolicy pertinent to the policy
language given to the function.  In this version of the code,
the policy language is described by a set of required tables that
must be relevant to the subpolicy, a set of prohibited tables that
may not be relevant to the subpolicy, and a set of output tables
that all rules must help define.

Change-Id: I0ed8f2035fc7026cfae39d1217659e1420a485ce