Fix bug in get_credentials
The method get_credentials had specific openstack arguments hardcoded which
prevented one from registering a driver with non-openstack config options.
This patch fixes this by removing this function and passing args['config']
directly into each driver. The check for required options no longer needs
to be done in the driver because it's done in the api-layer and knows the
options from the get_datasource_info() method. There are already tests that
cover this segment of code.

Change-Id: I92dd206fa43e34241e1836b534ac1dc72385b6ac