Add datasource synchronizer
The datasource synchronizer reads from the configuration database on a
periodic basis and checks whether the active datasources match the
configuration in the database.  If the two do to match, the
synchronizer changes the active datasources to match the database.

By default, datasource synchronization is off.  To enable it, add a
line like this
   datasource_sync_period = 60
to the congress.conf configuration file.

This change includes a change to the datasource status API semantics.
The previous version of the status handler first checks that the
datasource exists in the database; if the datasource does not the
handler responds to the client with and error because the running
state in the server always matches the database content because there
can be just one Congress server modifying the database.  With the
basic high-availability solution, one congress server replica can
modify the database without a second replica knowing, so the running
state of the datasources can become out of sync with the database.

The new version of the datasource status API call returns the status
of the datasource if the datasource is running in the server, even if
the database does not contain the datasource.

This change also makes the list-datasources API call update the
running datasources to match the database after fetching the
datasource configuration from the database.

This also adds a tempest test to verify that datasource
synchronization works.

Implements blueprint: basic-high-availability
Change-Id: I732c965c616f0f11d4d30f7f90b0cce980f9342d